Change the way custom modules are referenced
Neither the example in Reamde.md nor the predefined plugins have constrcturs. Thus just copying one of the predefined plugins and requiring them in the way as described in readme doesn't work. Therefore you need to explicitely pass a conrete function to PluginRunner.js. 

However imho a better solution would be to update the predefined modules.